requirements.txt
amqp==5.0.6
art==5.3
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
json-rpc==1.13.0
kombu==5.1.0
numpy==1.21.2
pandas==1.3.2
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pycm==3.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
sentry-sdk==1.3.1
six==1.16.0
sqlparse==0.4.1
torch==1.9.0
typing-extensions==3.10.0.2
urllib3==1.26.6
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
